+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to aws organization formation will be documented in this fil
 - Fixed bug with OU structures that go three levels deep and/or multiple times the same ou name
 - Added Support for `!Join` expressions in task files
 - Added support for `--organization-file` option on `print-stacks` command
+- Init command adds RootEmail to MasterAccount in generated template
+- Fixed bug where `!CopyValue` causes process to hang
 
 **version 0.9.6**
 - Fixed issue where `perform-task` parameters are passed to `update-stacks` task when no parameters where specified on task.